% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_synthetics}
\alias{add_synthetics}
\title{Add synthetic instruments to instrument list}
\usage{
add_synthetics(inst_list_raw, settings)
}
\arguments{
\item{inst_list_raw}{an object of type \code{\link{rh_instrument_list}}}

\item{settings}{see \code{\link{rh_new_curve}}}
}
\value{
an object of type \code{\link{rh_instrument_list}}
}
\description{
Add synthetic instruments to instrument list
}
\details{
the action performed is based on the class of inst_list_raw:
  rh_snt_ois. When this class is not assigned to inst_list_raw, 
  no synthetic instruments are added.
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "ois",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)

  settings$discount_curve <- "curve_ois"
  inst_list_ibor_wo <- rh_new_instrument_list(inst_df_ibor, settings)
  inst_list_ibor_snt <- add_synthetics(inst_list_ibor_wo, settings)
}
