% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark a curve to a reference}
\usage{
benchmark(curve, bench_zeros = NULL, bench_dfs = NULL, pillars = -1)
}
\arguments{
\item{curve}{a bootstrapped curve (created by \code{\link{bootstrap}})}

\item{bench_zeros}{a vector of benchmarking zero rates (can be NULL)}

\item{bench_dfs}{a vector of benchmarking discount factors (can be NULL)}

\item{pillars}{the indices for which the benchmarks apply.}
}
\value{
an object with fields zeros and dfs, containing the differences for the zero rates and the discount factors respectively
}
\description{
Benchmark a curve to a reference
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)
  benchmark(curve_ois, 
            bench_zeros = c(0, 0, 0), 
            bench_dfs = c(1, 1, 1),
            pillars = c(2, 5, 8))
}
