% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a yield curve}
\usage{
bootstrap(curve, inst_list)
}
\arguments{
\item{curve}{an object of class \code{\link{rh_curve}}, created by \code{\link{rh_new_curve}}}

\item{inst_list}{an object of class \code{\link{rh_instrument_list}}, created by \code{\link{rh_new_instrument_list}}}
}
\value{
a bootstrapped curve (class \code{\link{rh_curve}})
}
\description{
Bootstrap a yield curve
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)

  settings$discount_curve <- "curve_ois"
  inst_list_ibor <- rh_new_instrument_list(inst_df_ibor, settings)
  curve_ibor <- bootstrap(rh_new_curve(settings), inst_list_ibor)
}
