% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_convexity_adjustment}
\alias{calc_convexity_adjustment}
\title{Calculate the convexity adjustment for futures}
\usage{
calc_convexity_adjustment(instrument)
}
\arguments{
\item{instrument}{an object of classes \code{\link{rh_instrument}} and `rh_future`; with additionally the fields `price` and `convexity_adj` set:
\describe{
  \item{price}{the quoted futures price}
  \item{convexity_adj}{a structure with info about the convexity adjustment:
    \describe{
      \item{mean_rev_speed}{the mean reversion speed}
      \item{volatility}{the volatility}
    }}
}}
}
\value{
The equivalent FRA rate of the future
}
\description{
Calculate the convexity adjustment for futures
}
\details{
the method followed is that of Kirikos & Novak: Convexity Conundrums
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   ccy = "EUR")
  inst_df <- list(quote = 0.03,
                  start_no = 6,
                  start_unit = "MO",
                  mat_no = 3,
                  mat_unit = "MO",
                  insttype = "future")
  future <- rh_new_instrument(inst_df, settings)
  future$price <- 96.9448
  future$convexity_adj <- list(mean_rev_speed = 0.025,
                               volatility = 0.05)
  calc_convexity_adjustment(future)
}
\seealso{
Other convexity adjustment: 
\code{\link{calc_inv_convexity_adjustment}()}
}
\concept{convexity adjustment}
