% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{daycount}
\alias{daycount}
\title{apply daycount convention to calculate year fraction}
\usage{
daycount(object, fromdate, todate)
}
\arguments{
\item{object}{object with a class to determine which convention to apply, currently the classes "rh_act365", "rh_act360" and "rh_30e360" are implemented}

\item{fromdate}{start date of the period}

\item{todate}{end date of the period}
}
\value{
a number, the year fraction
}
\description{
apply daycount convention to calculate year fraction
}
\examples{
  o_act365 <- structure(list(), class = "rh_act365")
  o_act360 <- structure(list(), class = "rh_act360")
  o_30e360 <- structure(list(), class = "rh_30e360")
  daycount(o_act365, as.Date("2022-02-22"), as.Date("2022-08-10"))
  daycount(o_act360, as.Date("2022-02-22"), as.Date("2022-08-10"))
  daycount(o_30e360, as.Date("2022-02-22"), as.Date("2022-08-10"))
}
\seealso{
Other daycount: 
\code{\link{daycount.rh_30e360}()},
\code{\link{daycount.rh_act360}()},
\code{\link{daycount.rh_act365}()}
}
\concept{daycount}
