% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-arithmetic.R
\name{daycount.rh_30e360}
\alias{daycount.rh_30e360}
\title{apply 30E/360 convention to calculate year fraction}
\usage{
\method{daycount}{rh_30e360}(object, fromdate, todate)
}
\arguments{
\item{object}{object with a class to determine which convention to apply, currently the classes "rh_act365", "rh_act360" and "rh_30e360" are implemented}

\item{fromdate}{start date of the period}

\item{todate}{end date of the period}
}
\description{
apply 30E/360 convention to calculate year fraction
}
\details{
year(todate) - year(fromdate)
   + 1/12 (month(todate) - month(fromdate))
   + 1/360 (day(todate) - day(fromdate)) \cr
   when the day = 31, it is corrected to 30.
}
\seealso{
Other daycount: 
\code{\link{daycount.rh_act360}()},
\code{\link{daycount.rh_act365}()},
\code{\link{daycount}()}
}
\concept{daycount}
