% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-arithmetic.R
\name{daycount.rh_act360}
\alias{daycount.rh_act360}
\title{apply ACT/360 convention to calculate year fraction}
\usage{
\method{daycount}{rh_act360}(object, fromdate, todate)
}
\arguments{
\item{object}{object with a class to determine which convention to apply, currently the classes "rh_act365", "rh_act360" and "rh_30e360" are implemented}

\item{fromdate}{start date of the period}

\item{todate}{end date of the period}
}
\description{
apply ACT/360 convention to calculate year fraction
}
\details{
the actual number of days in the period, divided by 360
}
\seealso{
Other daycount: 
\code{\link{daycount.rh_30e360}()},
\code{\link{daycount.rh_act365}()},
\code{\link{daycount}()}
}
\concept{daycount}
