% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fixing}
\alias{fixing}
\title{Fixing: fixes a quote for an instrument based on a curve}
\usage{
fixing(instrument, curve)
}
\arguments{
\item{instrument}{a structure of class "rh_instrument"}

\item{curve}{a structure of class "rh_curve"}
}
\value{
the interest rate quote of the instrument according to the curve
}
\description{
Fixing: fixes a quote for an instrument based on a curve
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)
  fixing(inst_list_ois[[5]], curve_ois)

  inst_df <- list(quote = 0.01,
                  mat_no = 1,
                  mat_unit = "WK",
                  insttype = "deposit")
  depo1w <- rh_new_instrument(inst_df, settings)
  fixing(depo1w, curve_ois)
}
