% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_insta_fwd}
\alias{get_insta_fwd}
\title{extract instanteneous forward rates from a curve}
\usage{
get_insta_fwd(curve, dates = NULL, times = NULL, pillars = NULL)
}
\arguments{
\item{curve}{the curve from which to get the data}

\item{dates}{the dates for which the data is requested (can be missing)}

\item{times}{the times for which the data is requested (can be missing)}

\item{pillars}{the index of the pillars for wich the data is requested (can be missing)}
}
\value{
the instantaneous forward rates as per the request
}
\description{
extract instanteneous forward rates from a curve
}
\details{
at most one of dates, times and pillars should be given \cr
  if dates, times and pillars are NULL, the function returns the
  instantaneous forward rates for all pillars \cr
  if pillars is not NULL, the instantaneous forward rates for these
  pillars are returned \cr
  if pillars is NULL and dates is not NULL, the instantaneous forward rates
  for these dates are returned \cr
  if dates and pillars are NULL and times is not NULL, this is used to
  determine the instantaneous forward rates
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)
  get_insta_fwd(curve_ois)
  get_insta_fwd(curve_ois, times = c(0.5, 1, 2, 5))
}
