% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_rh_param}
\alias{get_rh_param}
\title{get RateHike parameter value}
\usage{
get_rh_param(param_name = NULL)
}
\arguments{
\item{param_name}{one of the parameters in \code{\link{rh_params}}}
}
\value{
the value of the requested parameter
}
\description{
get RateHike parameter value
}
\details{
if called without argument, all RateHike parameters are returned.
}
\examples{
   get_rh_param("root_find_tol")
   get_rh_param()
}
\seealso{
Other rh_params: 
\code{\link{set_rh_param}()}
}
\concept{rh_params}
