% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{graph}
\alias{graph}
\title{plot a yield curve}
\usage{
graph(curve, t = NULL, max_t = NULL, n_points = NULL, fwd_period = "6M")
}
\arguments{
\item{curve}{an object of class \code{\link{rh_curve}}}

\item{t}{a vector of timepoints}

\item{max_t}{the maximum time value to be plotted}

\item{n_points}{the number of points to be plotted}

\item{fwd_period}{the discrete forward period to apply}
}
\value{
a plot of the yield curve, including the zero rates, the discount factors, the instantaneous forward rates and the discrete forward rates
}
\description{
plot a yield curve
}
\details{
if t is provided, that is used, else a t is constructed
  based on max_t and n_points. \cr
  the default value for max_t is the last point on the curve, the default value for n_points is 500.
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)
  graph(curve_ois)
  graph(curve_ois, max_t = 2)
}
