% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inst_df_ibor}
\alias{inst_df_ibor}
\title{(dummy) input data for IBOR curve in dataframe format}
\format{
\describe{
  \item{quote}{market quote}
  \item{mat_no}{numeric component of maturity}
  \item{mat_unit}{unit for maturity (YR, MO)}
  \item{insttype}{instrument type (deposit, future, irs)}
  \item{start_no}{numeric component of start delay, for forward starting instruments}
  \item{start_unit}{unit for start delay (MO), for futures}
  \item{tenor_m}{tenor, in months, for irs}
}
}
\usage{
inst_df_ibor
}
\description{
A data frame with 29 observations and 7 variables.
Dummy data, represents inputs for the EURIBOR 3 months curve.
}
\details{
This dataframe can be used to run some examples in the package,
see for example \code{\link{bootstrap}}
}
\keyword{datasets}
