% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inst_df_ois}
\alias{inst_df_ois}
\title{(dummy) input data for OIS curve in dataframe format}
\format{
\describe{
  \item{quote}{market quote}
  \item{mat_no}{numeric component of maturity}
  \item{mat_unit}{unit for maturity (YR, MO, WK, DY)}
  \item{insttype}{instrument type (deposit, ois)}
}
}
\usage{
inst_df_ois
}
\description{
A data frame with 33 observations and 4 variables. 
Dummy data, represents inputs for the ESTR OIS curve.
}
\details{
This dataframe can be used to run some examples in the package,
see for example \code{\link{rh_new_instrument_list}}
}
\keyword{datasets}
