% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{leave_one_out}
\alias{leave_one_out}
\title{Leave-One-Out test for bootstrapping}
\usage{
leave_one_out(inst_list, settings, indices)
}
\arguments{
\item{inst_list}{an object of class \code{\link{rh_instrument_list}}, created by \code{\link{rh_new_instrument_list}}}

\item{settings}{see \code{\link{rh_new_curve}}}

\item{indices}{an index, or a vector of indices, of instruments to leave out}
}
\value{
an object with field diff, containing the differences between the fixings and the quotes for the instruments in inst_list
}
\description{
Leave-One-Out test for bootstrapping
}
\details{
This test bootstraps a curve on the full instrument list, and again on
a new instrument list with one instrument removed. The difference of 
the discount factor for the removed instrument is reported.
The argument `indices` can be a vector, in which case each index in the
vector is treated separately.
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  leave_one_out(inst_list_ois, settings, c(5, 10, 15))
}
