% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateHike-package.R
\docType{package}
\name{RateHike-package}
\alias{RateHike}
\alias{RateHike-package}
\title{RateHike: Ramp Up the Quality of Your Interest Rate Curves}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

RateHike is a package designed to bootstrap interest rate curves in a multi-curve framework, in particular IBOR curves and OIS curves. It was created as part of a master's thesis project, check out the vignette for more details. This thesis contains all the details on the implementation.
}
\section{Getting Started}{

The best start is to browse the documentation of the main objects: 
\code{\link{rh_curve}}, \code{\link{rh_instrument}} and 
\code{\link{rh_instrument_list}} or of the main function:
\code{\link{bootstrap}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spuddie/RateHike}
  \item Report bugs at \url{https://github.com/spuddie/RateHike/issues}
}

}
\author{
\strong{Maintainer}: Steff Helsen \email{RateHike@prophylaxis.be}

}
\keyword{internal}
