% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{re_anchor}
\alias{re_anchor}
\title{Re-anchor a curve}
\usage{
re_anchor(curve, depo_sn)
}
\arguments{
\item{curve}{an object of class \code{\link{rh_curve}}}

\item{depo_sn}{the spot-next deposit (only used for "ontn")}
}
\value{
an object of class \code{\link{rh_curve}}
}
\description{
Re-anchor a curve
}
\details{
the action performed depends on the class of curve; for the classes
  rh_anchor_date, rh_anchor_ontn and rh_anchor_11 a specific
  re-anchoring takes place; if none of these classes are assigned
  the curve is returned unmodified.
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "ontn",
                   snt = "ois",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois_2d <- bootstrap(rh_new_curve(settings), inst_list_ois)
  curve_ois_0d <- re_anchor(curve_ois_2d, inst_list_ois[[1]])
}
