\name{rh_instrument_list}
\alias{rh_instrument_list}
\title{Object of class rh_instrument_list}
\description{A list of financial instruments. Available instrument
  types are: deposit, ois, fra, future and irs}
\details{
  a list, of which the elements are objects of type
  \code{\link{rh_instrument}}. Created by
  \code{\link{rh_new_instrument_list}} or
  \code{\link{add_synthetics}}.
}
