% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-objects.R
\name{rh_new_instrument_list}
\alias{rh_new_instrument_list}
\title{creates an instrument list}
\usage{
rh_new_instrument_list(inst_df, settings)
}
\arguments{
\item{inst_df}{a data frame with the following fields:
\describe{
  \item{quote}{market quote}
  \item{mat_no}{numeric component of maturity}
  \item{mat_unit}{unit for maturity (YR, MO, WK, DY)}
  \item{insttype}{instrument type (deposit, ois, fra, future, irs)}
  \item{start_no}{numeric component of start delay, for forward starting instruments}
  \item{start_unit}{unit for start delay (YR, MO, WK, DY), for forward starting instruments}
  \item{tenor_m}{tenor, in months, for irs}
}}

\item{settings}{structure with the following fields:
\describe{
  \item{today_date}{today's date yyyy-mm-dd}
  \item{ccy}{the currency (currently only EUR and GBP are supported}
  \item{snt}{which synthetic instruments should be added? (no, ois)}
  \item{fut_ca_mrs}{optional; the mean reversion speed for the futures convexity adjustment}
  \item{fut_ca_vol}{optional; the volatility for the futures convexity adjustment}
}}
}
\value{
an object of class \code{\link{rh_instrument_list}}
}
\description{
creates an instrument list
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   ccy = "EUR",
                   snt = "no")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
}
