% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{roundtrip}
\alias{roundtrip}
\title{Round-trip test for a bootstrapped curve}
\usage{
roundtrip(curve, inst_list)
}
\arguments{
\item{curve}{a bootstrapped curve (created by \code{\link{bootstrap}})}

\item{inst_list}{an object of class \code{\link{rh_instrument_list}}, created by \code{\link{rh_new_instrument_list}}}
}
\value{
an object with field diff, containing the differences between the fixings and the quotes for the instruments in inst_list
}
\description{
Round-trip test for a bootstrapped curve
}
\details{
compare the fixings of an instrument list to the quotes
  when used on the instrument list that was used to bootstrap the curve, this is called a round-trip test.
}
\examples{
  settings <- list(today_date = "2022-11-30",
                   daycount = "act365",
                   interp_what = "df",
                   interp_how = "linear",
                   anchor = "no",
                   snt = "no",
                   ccy = "EUR")
  inst_list_ois <- rh_new_instrument_list(inst_df_ois, settings)
  curve_ois <- bootstrap(rh_new_curve(settings), inst_list_ois)
  roundtrip(curve_ois, inst_list_ois)
}
