% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_rh_param}
\alias{set_rh_param}
\title{set RateHike parameter value}
\usage{
set_rh_param(param_name, value)
}
\arguments{
\item{param_name}{one of the parameters in \code{\link{rh_params}}}

\item{value}{the new value of the parameter}
}
\value{
this function sets the requested parameter; it invisibly returns the previous version of all RateHike paramters
}
\description{
set RateHike parameter value
}
\examples{
   set_rh_param("root_find_tol", 1e-9)
}
\seealso{
Other rh_params: 
\code{\link{get_rh_param}()}
}
\concept{rh_params}
